+++ src/glUtilities/TextureBuffer.h
@@ -8,11 +8,11 @@ namespace gl {
         i32 m_size, m_usage, m_internalFormat;
 
     public:
-        TextureBuffer(const void *data, i32 size, i32 usage, i32 internalFormat);
+        TextureBuffer(const void *data, i64 size, i32 usage, i32 internalFormat);
         ~TextureBuffer();
 
-        void setBuffer(const void *data, i32 size, i32 usage = 0, i32 internalFormat = 0);
-        void updateBuffer(const void *data, i32 offset, i32 size);
+        void setBuffer(const void *data, i64 size, i32 usage = 0, i32 internalFormat = 0);
+        void updateBuffer(const void *data, i32 offset, i64 size);
 
         void bind(u32 slot) const;
         void unbind() const;
